refactor: reorganize value module
